requirements.txt
mcp>=1.0.0
mysql-connector-python>=8.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0